requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
autopep8==2.3.1
backcall==0.2.0
black==24.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
DateTime==5.5
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.1.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.6.1
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy==1.14.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.0
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tomli==2.0.1
torch==2.2.2
tornado==6.4.2
traitlets==5.14.3
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
zope.interface==7.2
